refactor: Extract useAccount hook from logs-collection
